[dev] [dev] Fix a cast error got from -Werror=conversion and -Wfatal-errors

+++ code/simulator/unit_tests/src/SimTest.cpp
@@ -80,7 +80,7 @@ TEST_F(SimTest, can_simulate_oscillating_cube)
     ASSERT_EQ(N, res.size());
     for (size_t i = 0 ; i < N ; ++i)
     {
-        const double t = i*dt;
+        const double t = (double)i*dt;
         ASSERT_EQ(13,                          res.at(i).x.size())         << "Time step: i=" << i;
         ASSERT_DOUBLE_EQ(t,                    res.at(i).t)                << "Time step: i=" << i;
         ASSERT_NEAR(0,                         res.at(i).x[XIDX(0)], eps)  << "Time step: i=" << i;
